requirements.txt
bqplot==0.12.43
googletrans==3.1.0a0
ipydatagrid==1.2.1
ipyfilechooser==0.6.0
ipython==8.12.3
ipywidgets==8.1.3
matplotlib==3.5.2
nltk==3.7
numpy==1.23.4
pandas==1.4.4
plotly==5.9.0
python_Levenshtein==0.25.1
seaborn==0.11.2
spacy==3.7.4
Unidecode==1.2.0
nbformat==5.10.4
